Implement live update with new blips in oldblips.html
Also implement remaning features missing compared to standard
index.html. Might be suitable now as replacement for the default page.

New blips are continuously fetched with long-polling. The graph is updated
if relevant (zoomed in at aggregation level "none", and showing the vicinity
of the new blip), and scrolled as needed.

Added showing in text form the last blip received, and the github banner.